Former-commit-id: e66c7a222a5cf225e36069062d3c9fe5182f00f9
Former-commit-id: 05d846a56ac81b9abb50fc0c88a082b19e5bc88c

+++ kernel/kernel-masterdata-service/src/main/java/io/mosip/kernel/masterdata/controller/RegistrationCenterController.java
@@ -147,7 +147,7 @@ public class RegistrationCenterController {
 	 */
 	@GetMapping("/v1.0/registrationcenters/{langcode}/{hierarchylevel}/{name}")
 	public RegistrationCenterResponseDto getRegistrationCenterByHierarchyLevelAndTextAndlangCode(
-			@PathVariable("langcode") String langCode, @PathVariable("hierarchylevel") Integer hierarchyLevel,
+			@PathVariable("langcode") String langCode, @PathVariable("hierarchylevel") Short hierarchyLevel,
 			@PathVariable("name") String name) {
 		return registrationCenterService.findRegistrationCenterByHierarchyLevelandTextAndLanguageCode(langCode,
 				hierarchyLevel, name);
@@ -222,7 +222,7 @@ public class RegistrationCenterController {
 	 */
 	@GetMapping("/v1.0/registrationcenters/{langcode}/{hierarchylevel}/names")
 	public RegistrationCenterResponseDto getRegistrationCenterByHierarchyLevelAndListTextAndlangCode(
-			@PathVariable("langcode") String langCode, @PathVariable("hierarchylevel") Integer hierarchyLevel,
+			@PathVariable("langcode") String langCode, @PathVariable("hierarchylevel") Short hierarchyLevel,
 			@RequestParam("name") List<String> names) {
 		return registrationCenterService.findRegistrationCenterByHierarchyLevelAndListTextAndlangCode(langCode,
 				hierarchyLevel, names);
